pyproject.toml
[tool.poetry]
name = "canpl-analytics"
version = "0.1.0"
description = ""
authors = ["athach"]
readme = "README.md"
packages = [{include = "canpl_analytics"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.3"
pandas = "^2.0.2"
sqlalchemy = "^2.0.15"
psycopg2-binary = "^2.9.6"
regex = "^2023.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
